Update lunametrics-youtube-v6.js
Modified the jQuery call so that only iframes containing youtube within the src will be checked. Also added some backup type checking on the SourceCheckA if statement as this was crashing the script when applied to iframe that didn't have youtube url as src. This is a little bit overkill but prevents the script taking out all subsequent JavaScript on the page if there is an error.